requirements.txt
apache-airflow-providers-cncf-kubernetes==5.1.1
apache-airflow-providers-snowflake==4.0.2
apache-airflow-providers-amazon==6.0.0
python-dotenv==0.21.1
